raspberry2/3: Use stable cloned MAC address on stable release
We have workaround using Network Manager's clone-mac-address=stable feature to
overcome the problem of Raspberry Pi not having a proper MAC address on our
Raspberry Pi 2 and 3 builds.

Raspberry Pi 2 and 3 now have a stable MAC address on recent kernels. This is
because the CPU id is used to auto-generate a MAC address. This happens at the
driver level. So, this workaround is not needed.

Further, there is bug in Network Manager that makes the MAC address oscillate.
See https://bugzilla.gnome.org/show_bug.cgi?id=793231 . cloned-mac-address can't
be used reliably. Restrict the workaround to just Debian stable based build.

Signed-off-by: Sunil Mohan Adapa <sunil@medhas.org>
Reviewed-by: James Valleroy <jvalleroy@mailbox.org>