Fix naming convention for virtualbox log file
The log file naming convention for virtualbox builds is the same as that for
non-vm builds for the same architecture. For example, amd64 and virtualbox-amd64
will produce log files with the same name which will get overwritten when
written to the same directory on the ftp server.

With this fix, a virtualbox build's log file will end with .vdi.log instead of
.log

This naming convention is consistent with the naming convention of the image.

Signed-off-by: Joseph Nuthalapati <njoseph@thoughtworks.com>
Reviewed-by: Sunil Mohan Adapa <sunil@medhas.org>