Fix issue leading to unbootable grub configuration
Due to a bug, probably in udev, /dev/disk/by-uuid/<uuid> link is not reliably
created after the creation of the filesystem. This leads to update-grub using
root=/dev/mapper/loop0p1 instead of root=UUID=<uuid> when creating grub.cfg.
This results in an unbootable image. Force udev events as a workaround.

Signed-off-by: Sunil Mohan Adapa <sunil@medhas.org>
Reviewed-by: James Valleroy <jvalleroy@mailbox.org>